Restructuring my project in just one root directory
