+++ src/main/java/com/googleclone/googleclone/GooglecloneApplication.java
@@ -1,13 +1,13 @@
-package com.searchengine.searchengine;
+package com.googleclone.googleclone;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SearchengineApplication {
+public class GooglecloneApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SearchengineApplication.class, args);
+		SpringApplication.run(GooglecloneApplication.class, args);
 	}
 
 }
