+++ src/main/java/com/googleclone/googleclone/ServletInitializer.java
@@ -1,4 +1,4 @@
-package com.searchengine.searchengine;
+package com.googleclone.googleclone;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
@@ -7,7 +7,7 @@ public class ServletInitializer extends SpringBootServletInitializer {
 
 	@Override
 	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(SearchengineApplication.class);
+		return application.sources(GooglecloneApplication.class);
 	}
 
 }
