+++ src/test/java/com/googleclone/googleclone/GooglecloneApplicationTests.java
@@ -1,10 +1,10 @@
-package com.searchengine.searchengine;
+package com.googleclone.googleclone;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class SearchengineApplicationTests {
+class GooglecloneApplicationTests {
 
 	@Test
 	void contextLoads() {
